Remove unused href attributes
The use of href attributes here was to circumvent a constraint of the
DSFR that disable links unless they have an explicite href attribute.
Edit the dsfr.min.css instead, until this is fixed upstream.

This fixes a bug where sorting was triggering the display of the full
list of repos, ignoring possible filters.

Thanks to L. Nussbaum for reporting this.